pkg: Remove function `get_forge_download`
As part of removing "forge" nomenclature from `pkg`, the function
`get_forge_download` was found to be very short (only 3 lines) and it was also
being used in exactly one location in pkg.m. Instead of renaming the file
to something without "forge", it is easier to inline its contents where
it was being called in pkg.m and delete the file instead.

https://octave.discourse.group/t/suggestions-for-clearer-name-besides-forge-to-use-with-pkg-command/6358/51

All entries refer to the directory scripts/pkg:

* pkg.m: Inline the contents of the `get_forge_download` function where it
  was called, rename some variables, reuse existing temporary directory,
  add a comment for future programmers.

* private/get_forge_download.m: Delete file.

* module.mk: Remove file from build system.