pkg: Replace 'eval' with JSON
The package metadata returned by packages.octave.org was recently changed
from 'eval' to 'jsondecode' instead. Link:
https://octave.discourse.group/t/strengthen-checks-for-package-parsing-code/6357/18

This changeset accommodates that change.

One new complication is that for packages that test themselves in CI for
certain build environments, their metadata uses a cell array instead of a
struct, which is what is returned for packages that do not use that specific
CI environment. This unavoidable inconsistency is generated server-side
and is worked around in 'pkg' code. Since the syntax to dig deeper into a
cell differs from that for a struct, a new private function has been created
to avoid code duplication.

All changes are in the directory 'scripts/pkg/private' unless marked otherwise:

* get_validated_pkg_list.m: Change URL for package metadata to the
one for the new JSON format, delete all old validation code (superseded).

* convert_possible_cell_to_struct.m: New file and function to work around
the cell/struct inconsistency described above.

* get_pkg_info.m: Update to use new format.

* search_packages.m: Update to use new format.

* scripts/pkg/module.mk: Add new file to build system.