Remove visibility attribute from function that is defined in header.
* liboctave/array/dim-vector.h (dim_vector::dim_max): Remove visibility
attribute from function that is defined inline in the header. It is not
necessary to export that function from a library because its definition will be
available in all compilation units that include this header anyway. Adding a
dllimport attribute to a function with definition is an error on Windows.