Check for undefined outputs in cellfun & arrayfun when nargout=0 (bug#66642).
* cellfun.cc (Fcellfun, Farrayfun): Set expected_nargout to nargout, except
when nargout is zero, in which case expected_nargout will be either 1, if the
first function call returns something, or zero otherwise.
Check for unexpected outputs when nargout=0 and the first function call did not
return anything, with UniformOutput=true.
Add tests with nargout=0. Add tests with varargout and undefined outputs.
Modify some tests to mark fails as regressions.